Create index.html
Simple, functional,voice timer-alarm clock. The first time you press the button or space bar, the microphone turns on, you have a few seconds to record the sound or note that will be played when the signal is triggered. Then you should press the second time and say the time required for the alarm or the minute after which the timer will be triggered. If you do not remove the waste, after a minute there will be a short reminder signal. To delete a reminder, you must click on the reminder line twice. In short, just speak into the microphone, the time is set without having to look for numbers, which saves time and is simply convenient.

Financial thanks are accepted to stimulate creativity